Refactor: removed extension error from util
